feature: update to new configuration
This commit is the first step to refactoring the application
to mimic the behavior of the official cli-synchronization
offering. The commandline options are removed in favor of a
json configuration file. The api-token is removed in favor of
a (untested) socket connection. Logging level and additional
features have been added to the configuration file.